Added 5 experimental results
Added experimental results for 5 models. All of the glove.6B... results are in their final form. I'm going to redo glove.42B.300d if I have time. Because some of the analogy files were too big with my batch size but I fixed that in the code so I should be able to run those now.